permission added to controller functions

+++ application/controllers/item.php
@@ -2,11 +2,27 @@
 
 class Item extends CI_Controller {
 
+    public function isAccessDenied($url) {
+        $perm = $this->session->userdata('permission');
+        if ($perm) {
+            $permission = explode(",", $perm);
+            if (in_array($url, $permission)) {
+                return FALSE;
+            }
+        }
+        return TRUE;
+    }
+    
     public function index() {
         $this->home();
     }
 
     public function home() {
+        if ($this->isAccessDenied("item"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('item_model');
         $this->load->model('itemcategory_model');
         $this->load->model('itemprice_model');
@@ -17,6 +33,11 @@ class Item extends CI_Controller {
     }
 
     public function add() {
+        if ($this->isAccessDenied("item/add"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('itemcategory_model');
         
         $data['main_content'] = "additem_form";
@@ -24,6 +45,11 @@ class Item extends CI_Controller {
     }
 
     public function add_db() {
+        if ($this->isAccessDenied("item/add_db"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('item_model');
         $this->item_model->add_item();
 
@@ -31,6 +57,11 @@ class Item extends CI_Controller {
     }
 
     public function update($item_id) {
+        if ($this->isAccessDenied("item/update"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('item_model');
         $this->load->model('itemcategory_model');
         
@@ -40,6 +71,11 @@ class Item extends CI_Controller {
     }
 
     public function update_db() {
+        if ($this->isAccessDenied("item/update_db"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('item_model');
         $this->item_model->update_item();
 
@@ -47,6 +83,11 @@ class Item extends CI_Controller {
     }
 
     public function delete($item_id) {
+        if ($this->isAccessDenied("item/delete"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('item_model');
         $this->item_model->delete_item(base64_decode(urldecode($item_id)));
