+++ application/controllers/stock.php
@@ -2,11 +2,27 @@
 
 class Stock extends CI_Controller {
 
+    public function isAccessDenied($url) {
+        $perm = $this->session->userdata('permission');
+        if ($perm) {
+            $permission = explode(",", $perm);
+            if (in_array($url, $permission)) {
+                return FALSE;
+            }
+        }
+        return TRUE;
+    }
+    
     public function index() {
         $this->home();
     }
 
     public function home() {
+        if ($this->isAccessDenied("stock"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('stock_model');
         $this->load->model('item_model');
         $this->load->model('itemcategory_model');
@@ -17,6 +33,11 @@ class Stock extends CI_Controller {
     }
 
     public function add($item_id) {
+        if ($this->isAccessDenied("stock/add"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('itemprice_model');
         
         $data['main_content'] = "addstockitem_form";
@@ -25,6 +46,11 @@ class Stock extends CI_Controller {
     }
     
     public function add_db() {
+        if ($this->isAccessDenied("stock/add_db"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('stock_model');
         $this->stock_model->add_stockitem();
         
@@ -32,6 +58,11 @@ class Stock extends CI_Controller {
     }
 
     public function update($item_id) {
+        if ($this->isAccessDenied("stock/update"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('stock_model');
         $this->load->model('itemprice_model');
         
@@ -41,13 +72,23 @@ class Stock extends CI_Controller {
     }
     
     public function update_db() {
+        if ($this->isAccessDenied("stock/update_db"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('stock_model');
         $this->stock_model->update_stockitem();
         
         redirect(base_url() . 'stock');
     }
     
-    public function delete_db($stock_id) {
+    public function delete($stock_id) {
+        if ($this->isAccessDenied("stock/delete"))
+        {
+            redirect(base_url());
+        }
+        
         $this->load->model('stock_model');
         $this->stock_model->delete_stockitem(base64_decode(urldecode($stock_id)));
